server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Kevin Hu <kevinhuwest@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
docker = "^5.0.3"
prisma = "^0.6.6"
strictyaml = "^1.6.1"
aiofiles = "^0.8.0"
python-multipart = "^0.0.5"
pydantic-yaml = {extras = ["pyyaml"], version = "^0.8.0"}
requests = "^2.28.1"
PyJWT = "^2.4.0"
cryptography = "^37.0.4"
PyYAML = "^6.0"
village-temporalio = "^0.1b2"
ujson = "^5.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
isort = "^5.10.1"
pyright = "^1.1.266"
pylint = "^2.14.5"
vulture = "^2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = false

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
profile = "black"